RobotLink: allow subclassing and optionally show as tree
Expands the list of link properties to optionally show the list as a tree of
links and/or joints.  The Style property allows user to choose style.  Default
is same as before (list of links in alphabetic order).

Add setLinkFactory() which allows a subclass of RobotLink and RobotJoint to be used.

Make friendlier to subclassing by changing some members from private to protected.