Refactor how data is passed between widgets when adding a display
Have tabs contained by AddDisplayDialog emit SelectionData whenever a new
selection is made.  This is class contains all of the knowledge needed to
select a plugin along with an optional topic and datatype.
AddDisplayDialog now saves the current selection information for each tab
and reloads it when the tab changes.