Change add dialog to use combo boxes for similar topics
In the display dialog, if a set of topics are subsets of one base topic
and are compatible with the same plugins, display the base topic along
with a combo box to select the subtopic.

For example, suppose camera data is being published on /image,
/image/compressed, and /image/theora and those topics are compatible with
the Camera and Image plugins.  Then the display dialog should display
/image as a topic with a combo box containing the entries "raw",
"compressed", and "theora" with the Camera and Image plugins below it.
Previously, each topic would have appeared as a separate entry.