Use a tree with combo boxes for selection by topic
In "Add display by topic dialog" topics are now displayed hierarchically.
If a topic is visualizable (i.e., at least one plugin it says it can use
it) all plugins that can visualize that topic are listed as children.  If
a plugin can view topics that are children of the visualizable topic, they
are displayed as dropdown options in a combo box of the plugin.