Smooth updates for map display plugin
Map displays previously only updated when receiving a message. This means that
if your fixed frame was base_link, the costmaps would not move appropriately
around the robot unless a message was received in order to update the transform
that should be applied to the scene. For global costmaps, this is a slow
update and for static maps, this never happened.

This fixes that by hooking into rviz' periodic call to continuously update the
transform to be applied to the scene.