Correct xmlDoc.Parse usage.
`MemoryResource` returns the buffer pointer and size but doesn't have the guarantee the buffer being null-terminated (which makes sense since `MemoryResource` could hold anything in binary format.)

However, xmlDoc.Parse() does require a null-terminated string, or a buffer size needs to be specified. Otherwise, when the parser reads the bytes beyond the actual buffer size, it will lead to unexpected behavior (In my observation, this issue will manifest as `XML_ERROR_PARSING_TEXT` error.) and a scale of 1.0 will be returned.

So, specify the buffer size when calling Parse() to make sure it only reads till the buffer end.