Add ability to set alpha channel value of Axes and TF displays
This adds a property named alpha to the Axes and TF displays that
changes the alpha channel value of the axes colors.

The updateAlpha method is added to Axes so that TF display can easily
change the default alpha value of the axes.

The default axes colors in the Axes object are changed to be fields
instead of static objects so that different axes objects can have
different alpha values.