ogre helpers point_cloud: fix bounding radius
The PointCloudRenderable returns an incorrect bounding radius when
the furthest corner of the box is not the minimum or maximum corner.
This results in incorrect culling of point cloud renderables in
certain viewing angles when zooming in on the cloud.

Fix this by finding the largest absolute value of each coordinate
to find the furthest corner from the origin. Then return the distance
to the origin as the bounding radius.