always forward keystrokes in display view to property editor
Behavior without this patch:
- Open RViz with any display added
- click on the right column in the display view, in the same row as the checkbox to enable/disable the display
=> pressing spacebar can toggle the checkbox
- click on the left column of the same row (the name of the display)
=> pressing spacebar has no effect

Additionally, a visible indicator of the current focus (not the current selection)
highlights the corresponding cell.

This patch modifies selection requests to always focus the right column (if available)
and achieve the first behavior consistently.

This allows to go through a long list of displays by keyboard and toggle them with spacebar.

Do not redirect index if it is marked as Editable.
This is not the case with the current PropertyTree,
but might prevent unintuitive behavior in the future.