PropertyTree: set custom SelectionModel only with valid model
This seems to be a corner-case one might actually call a bug in qt5.
the `setSelectionModel(new PropertySelectionModel(nullptr))` call
complains that the SelectionModel and the PropertyTreeView use different
models, although both are set to nullptr in user code.

Qt actually sets the model of QAbstractItemView to
the dummy `QAbstractItemModelPrivate::staticEmptyModel`, but the
SelectionModel is actually initialized with `nullptr`.