SplitterHandle: Consider scrollbar width
If there is a scrollbar visible, the visible width of the viewport is reduced,
which should be considered when computing the width of the second column.
Without this, the viewport will scroll horizontally to bring the 2nd column into view,
thus hiding the unfold arrows in the first column. Fixes #1545.