Use threaded callback queue for PointCloud displays
PointCloudCommon was defining its own callback queue and spinner
to allow for asynchronous processing of point cloud data.
However, as most PointCloud displays just inherit MessageFilterDisplay,
which is using the default update_nh_, eventually the default synchronous
callback queue was used.
(Although the callback queue was configured in the constructor to point
to PointCloudCommon's private queue, Display::initialize() was resetting
the queue back to the default.)