Fixup assimp loading
This is a fixup of 7085a1402ec3b002dd3b7c3d86296d80f8d04d3f.
Obviously, assimp's root node is only sometimes, but not always applying a rotation
to y-up orientation. This fixup, drops the rotation part of the root transform,
but keeps the scaling (and position offset).